+++ ios/Classes/InstabugFlutterPlugin.m
@@ -6,8 +6,9 @@
 @implementation InstabugFlutterPlugin
 
 
+FlutterMethodChannel* channel;
 + (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  FlutterMethodChannel* channel = [FlutterMethodChannel
+  channel = [FlutterMethodChannel
       methodChannelWithName:@"instabug_flutter"
             binaryMessenger:[registrar messenger]];
   InstabugFlutterPlugin* instance = [[InstabugFlutterPlugin alloc] init];
@@ -357,6 +358,57 @@
     NSDictionary *constants = [self constants];
     NSInteger welcomeMode = ((NSNumber *) constants[welcomeMessageMode]).integerValue;
     [Instabug setWelcomeMessageMode:welcomeMode];
+}
+
+/**
+  * Enables and disables manual invocation and prompt options for bug and feedback.
+  * @param {boolean} isEnabled
+  */
++ (void)setBugReportingEnabled:(NSNumber *)isEnabled {
+   BOOL boolValue = [isEnabled boolValue];
+   IBGBugReporting.enabled = boolValue;
+}
+
+/**
+  * Sets a block of code to be executed just before the SDK's UI is presented.
+  * This block is executed on the UI thread. Could be used for performing any
+  * UI changes before the SDK's UI is shown.
+  */
++ (void)setOnInvokeCallback {
+   IBGBugReporting.willInvokeHandler = ^{
+             [channel invokeMethod:@"onInvokeCallback" arguments:nil];
+        };
+}
+
+/**
+  * Sets a block of code to be executed right after the SDK's UI is dismissed.
+  * This block is executed on the UI thread. Could be used for performing any
+  * UI changes after the SDK's UI is dismissed.
+  */
++ (void)setOnDismissCallback {
+   IBGBugReporting.didDismissHandler = ^(IBGDismissType dismissType, IBGReportType reportType) {            
+            //parse dismiss type enum
+            NSString* dismissTypeString;
+            if (dismissType == IBGDismissTypeCancel) {
+                dismissTypeString = @"CANCEL";
+            } else if (dismissType == IBGDismissTypeSubmit) {
+                dismissTypeString = @"SUBMIT";
+            } else if (dismissType == IBGDismissTypeAddAttachment) {
+                dismissTypeString = @"ADD_ATTACHMENT";
+            }
+            //parse report type enum
+            NSString* reportTypeString;
+            if (reportType == IBGReportTypeBug) {
+                reportTypeString = @"bug";
+            } else if (reportType == IBGReportTypeFeedback) {
+                reportTypeString = @"feedback";
+            } else {
+                reportTypeString = @"other";
+            }
+            NSDictionary *result = @{ @"dismissType": dismissTypeString,
+                                     @"reportType": reportTypeString};
+             [channel invokeMethod:@"onDismissCallback" arguments:result];
+        };
 }
 
 + (NSDictionary *)constants {